Moved Db content to Storage namespace
This was necessary since those models will only change with references
to the storage requirements such as the choice of database technology to
use etc.

We have also bundled the change that generates The Twitter video Model
from Twitter Notification model.

Upnext is the work around how the data gets stored in an easy to
retrieve manner, that is, marrying caching with DB